Extract all types to files
