+++ Megasware128.HalfLifeLauncher/ConfigArgument.cs
@@ -1,0 +1,1 @@
+enum ConfigArgument { Get, Set }
