+++ Megasware128.HalfLifeLauncher/ConfigOptions.cs
@@ -1,0 +1,5 @@
+class ConfigOptions
+{
+    public string HalfLifeDirectory { get; set; }
+    public string SteamDirectory { get; set; }
+}
