+++ Megasware128.HalfLifeLauncher/Program.cs
@@ -4,10 +4,6 @@ using System.CommandLine.Completions;
 using System.CommandLine.Hosting;
 using System.CommandLine.NamingConventionBinder;
 using System.CommandLine.Parsing;
-using System.Text;
-using Microsoft.Extensions.Configuration.Ini;
-using Microsoft.Extensions.FileProviders;
-using Microsoft.Extensions.Options;
 
 var settingsFile = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "Megasware128", "HalfLifeLauncher", "settings.ini");
 
@@ -115,114 +111,3 @@ return await new CommandLineBuilder(command)
     .UseDefaults()
     .Build()
     .InvokeAsync(args);
-
-class ConfigOptions
-{
-    public string HalfLifeDirectory { get; set; }
-    public string SteamDirectory { get; set; }
-}
-
-class ConfigService : BackgroundService
-{
-    private readonly ConfigOptions _options;
-    private readonly IConfiguration _config;
-
-    public ConfigService(IOptions<ConfigOptions> options, IConfiguration config)
-    {
-        _options = options.Value;
-        _config = config;
-    }
-
-    protected override Task ExecuteAsync(CancellationToken stoppingToken)
-    {
-        _config["HalfLifeDirectory"] = _options.HalfLifeDirectory;
-        _config["SteamDirectory"] = _options.SteamDirectory;
-
-        return Task.CompletedTask;
-    }
-}
-
-enum ConfigArgument { Get, Set }
-
-class SettingsFileConfigProvider : IniConfigurationProvider
-{
-    private FileStream? _stream;
-
-    public SettingsFileConfigProvider(IniConfigurationSource source) : base(source)
-    {
-    }
-
-    public override void Set(string key, string value)
-    {
-        base.Set(key, value);
-
-        if (_stream is null)
-        {
-            var path = Source.FileProvider.GetFileInfo(Source.Path).PhysicalPath;
-            Directory.CreateDirectory(Path.GetDirectoryName(path)!);
-            _stream = File.OpenWrite(path);
-            _stream.SetLength(0);
-            _stream.Seek(0, SeekOrigin.Begin);
-        }
-
-        _stream.Write(Encoding.UTF8.GetBytes($"{key}={value}\n"));
-    }
-
-    protected override void Dispose(bool disposing)
-    {
-        base.Dispose(disposing);
-
-        _stream?.Dispose();
-    }
-}
-
-class SettingsFileConfigurationSource : IniConfigurationSource
-{
-    public override IConfigurationProvider Build(IConfigurationBuilder builder)
-    {
-        EnsureDefaults(builder);
-        return new SettingsFileConfigProvider(this);
-    }
-}
-
-static class SettingsConfigurationExtensions
-{
-    public static IConfigurationBuilder AddSettingsFile(this IConfigurationBuilder builder, string path)
-    {
-        return AddSettingsFile(builder, provider: null, path: path, optional: false, reloadOnChange: false);
-    }
-
-    public static IConfigurationBuilder AddSettingsFile(this IConfigurationBuilder builder, string path, bool optional)
-    {
-        return AddSettingsFile(builder, provider: null, path: path, optional: optional, reloadOnChange: false);
-    }
-
-    public static IConfigurationBuilder AddSettingsFile(this IConfigurationBuilder builder, string path, bool optional, bool reloadOnChange)
-    {
-        return AddSettingsFile(builder, provider: null, path: path, optional: optional, reloadOnChange: reloadOnChange);
-    }
-
-    public static IConfigurationBuilder AddSettingsFile(this IConfigurationBuilder builder, IFileProvider? provider, string path, bool optional, bool reloadOnChange)
-    {
-        if (builder == null)
-        {
-            throw new ArgumentNullException(nameof(builder));
-        }
-        if (string.IsNullOrEmpty(path))
-        {
-            throw new ArgumentException("Path cannot be null or empty.", nameof(path));
-        }
-
-        return builder.AddIniFile(s =>
-        {
-            s.FileProvider = provider;
-            s.Path = path;
-            s.Optional = optional;
-            s.ReloadOnChange = reloadOnChange;
-            s.ResolveFileProvider();
-        });
-    }
-
-    public static IConfigurationBuilder AddSettingsFile(this IConfigurationBuilder builder, Action<IniConfigurationSource> configureSource)
-        => builder.Add(configureSource);
-}
